Store the absolute home dir in globals, subtract heading locally.

+++ MinimOsd-Extra_Copter_Pre release_Beta/OSD_Vars.h
@@ -118,7 +118,7 @@ static float        osd_home_lon = 0;               // home longitude
 static float        osd_home_alt = 0; 
 static float        osd_alt_to_home = 0; 
 static long         osd_home_distance = 0;          // distance from home
-static uint8_t      osd_home_direction = 1;         // Arrow direction pointing to home (1-16 to CW loop)
+static int          osd_home_direction = 1;         // direction to home
 
 static float        osd_pitch = 0;                  // pitch from DCM
 static float        osd_roll = 0;                   // roll from DCM
